fix: rename powerworld/aux.rs, a Windows-reserved filename that broke release checkout
AUX is a reserved DOS device name, so refs/tags/v0.2.0 failed to check
out on the x86_64-w64-mingw32 release job (git: invalid path). The module
is powerworld::auxiliary now, and the encoding CI job gains a guard that
rejects any path Windows cannot check out (reserved stems, trailing
dot/space, <>:"\|?*) so the breakage surfaces in PR CI, not at tag time.
