+++ powerio/src/format/powerworld/auxiliary.rs
